refactor: split App into TheHeader/TheFooter and Home view; add router
