+++ src/components/TheFooter.vue
@@ -1,0 +1,21 @@
+<template>
+  <footer class="app-footer">
+    <div>© HTMLScore</div>
+  </footer>
+</template>
+
+<script>
+export default {
+  name: 'TheFooter',
+}
+</script>
+
+<style scoped>
+.app-footer {
+  background: #eee;
+  padding: 8px 20px;
+  text-align: center;
+  margin-top: 20px;
+  font-size: 0.9rem;
+}
+</style>
