+++ src/components/TheHeader.vue
@@ -1,0 +1,23 @@
+<template>
+  <header class="app-header">
+    <h1>HTML Score</h1>
+  </header>
+</template>
+
+<script>
+export default {
+  name: 'TheHeader',
+}
+</script>
+
+<style scoped>
+.app-header {
+  background: #3c3c3c;
+  color: white;
+  padding: 10px 20px;
+}
+.app-header h1 {
+  margin: 0;
+  font-size: 1.1rem;
+}
+</style>
